+++ previous-versions/commerce/src/sdk/sdk.ts
@@ -62,9 +62,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.37.0";
-    genVersion = "2.159.2";
-    userAgent = "speakeasy-sdk/typescript 0.37.0 2.159.2 3.0.0 @codat/commerce";
+    sdkVersion = "0.38.0";
+    genVersion = "2.169.0";
+    userAgent = "speakeasy-sdk/typescript 0.38.0 2.169.0 3.0.0 @codat/commerce";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);